Machine leaning/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
ydata-profiling
optuna
lazypredict
mlflow
streamlit
joblib
pyyaml
imbalanced-learn
protobuf<=3.20.3
kaleido